added block_id to message block
